Modularized app.js
